Extract utils/html
