+++ web/ui/react-app/src/utils/html.test.ts
@@ -1,0 +1,9 @@
+import { escapeHTML } from './html';
+
+describe('escapeHTML', (): void => {
+  it('escapes html sequences', () => {
+    expect(escapeHTML(`<strong>'example'&"another/example"</strong>`)).toEqual(
+      '&lt;strong&gt;&#39;example&#39;&amp;&quot;another&#x2F;example&quot;&lt;&#x2F;strong&gt;'
+    );
+  });
+});
